debian: Load XKB keyboard settings only for real keyboards
On the bitland laptop the udev rule filter ENV{ID_INPUT_KEY}=="?*" is
matching at least 3 different devices, only one being the real keyboard.

This is causing XKBMODEL to be set to "pc105" during FBE instead of
"chromebook" and making the media keys not functioning on the first
boot.

See https://phabricator.endlessm.com/T21566#597784 for a more extensive
analysis of this problem.

We fix this problem narrowing down the import filter only on real
keyboards marked with ID_INPUT_KEYBOARD by systemd.

Signed-off-by: Carlo Caione <carlo@endlessm.com>