Revert "composite: fix page flip for fullscreen application"
This reverts commit fe4d141fc7fe2d7ec13ce5093a8d746c406115dd.

I'm assuming this was an attempt to automatically unredirect fullscreen
windows.

It doesn't work - LibreOffice full screen presentations don't appear
at all.

In this case, the added condition causes the window not to be redirected,
but it does not perform the actions which seem necessary for unredirecting
it (compSetParentPixmap, compRestoreWindow), so it does not appear at all.

Revert this problematic patch.

If an app does want to have an unredirected full screen window, it can
set an X property to that effect. mutter also has its own logic
to automatically unredirect fullscreen windows when the conditions are
right.

https://phabricator.endlessm.com/T22327