Use regular xserver-common package
xwayland was recently added to Endless. It depends directly on
xserver-common, which results in xserver-common and
xserver-common-rockchip both being included, causing ostree build failures
due to conflicting files.

xserver-common is "distant" enough from the codebase that the simplest
way to deal with this is just to use the "official" xserver-common
package rather than building our own version here.